[230704] 지도 초기화까지, 다이얼로그 생성

+++ gather-e_android/app/src/main/java/com/kosa/gather_e/model/entity/gather/GatherEntity.kt
@@ -7,7 +7,7 @@ data class GatherEntity(
     var gatherSeq: Int?,
     var categorySeq: Int,
     var categoryName: String,
-    var gatherCreator: Int,
+    var gatherCreator: Long,
     var creatorName: String?,
     var creatorImgUrl: String?,
     var gatherTitle: String,
